requirements.txt
djangorestframework-simplejwt==5.2.2
django==4.1.3
djangorestframework==3.14.0
xlwt~=1.3.0
xhtml2pdf~=0.2.8